Make gaphor.diagram.diagramtools a package
