+++ gaphor/diagram/diagramtools/placement.py
@@ -1,26 +1,13 @@
-"""Tools for handling items on the canvas.
-
-Although Gaphas has quite a few useful tools, some tools need to be extended:
- - PlacementTool: should perform undo
- - HandleTool: should support adapter based connection protocol
- - TextEditTool: should support adapter based edit protocol
-"""
-
 import logging
 
 from gaphas.aspect.connector import Connector as ConnectorAspect
 from gaphas.aspect.connector import ItemConnector
-from gaphas.aspect.move import Move as InMotionAspect
-from gaphas.guide import GuidedItemMove
-from gaphas.tool import hover_tool, item_tool, rubberband_tool
 from gaphas.tool.itemtool import item_at_point
-from gi.repository import Gdk
 
-from gaphor.core import Transaction, transactional
+from gaphor.core import transactional
 from gaphor.diagram.connectors import Connector
 from gaphor.diagram.event import DiagramItemPlaced
 from gaphor.diagram.grouping import Group
-from gaphor.diagram.inlineeditors import InlineEditor
 from gaphor.diagram.presentation import Presentation
 
 log = logging.getLogger(__name__)
@@ -119,33 +106,6 @@ class DisconnectHandle:
             if cinfo:
                 adapter = Connector(cinfo.connected, item)
                 adapter.disconnect(handle)
-
-
-class TextEditTool:
-    """Text edit tool.
-
-    Allows for elements that can adapt to the IEditable interface to be
-    edited.
-    """
-
-    def on_key_press(self, event):
-        view = self.view
-        item = view.selection.hovered_item
-        if (
-            item
-            and event.type == Gdk.EventType.KEY_PRESS
-            and event.key.keyval == Gdk.KEY_F2
-        ):
-            return InlineEditor(item, view)
-        return False
-
-    def on_double_click(self, event):
-        view = self.view
-        item = view.selection.hovered_item
-        if item:
-            x, y = view.get_matrix_v2i(item).transform_point(event.x, event.y)
-            return InlineEditor(item, view, (x, y))
-        return False
 
 
 class PlacementTool:
@@ -269,80 +229,3 @@ class PlacementTool:
             self._parent = None
             view.selection.set_dropzone_item(None)
         return item
-
-
-@InMotionAspect.register(Presentation)
-class DropZoneMove(GuidedItemMove):
-    def move(self, pos):
-        """Move the item.
-
-        x and y are in view coordinates.
-        """
-        super().move(pos)
-        item = self.item
-        view = self.view
-        x, y = pos
-
-        current_parent = view.canvas.get_parent(item)
-        over_item = item_at_point(view, (x, y), selected=False)
-
-        if not over_item:
-            view.selection.set_dropzone_item(None)
-            return
-
-        if current_parent and not over_item:
-            # are we going to remove from parent?
-            group = Group(current_parent, item)
-            if group:
-                view.selection.set_dropzone_item(current_parent)
-                current_parent.request_update(matrix=False)
-
-        if over_item:
-            # are we going to add to parent?
-            group = Group(over_item, item)
-            if group and group.can_contain():
-                view.selection.set_dropzone_item(over_item)
-                over_item.request_update(matrix=False)
-
-    def stop_move(self):
-        """Motion stops: drop!"""
-        super().stop_move()
-        item = self.item
-        view = self.view
-        canvas = view.canvas
-        old_parent = view.canvas.get_parent(item)
-        new_parent = view.selection.dropzone_item
-        try:
-
-            if new_parent is old_parent:
-                if old_parent is not None:
-                    old_parent.request_update(matrix=False)
-                return
-
-            if old_parent:
-                canvas.reparent(item, None)
-
-                adapter = Group(old_parent, item)
-                if adapter:
-                    adapter.ungroup()
-
-                old_parent.request_update()
-
-            if new_parent:
-                canvas.reparent(item, new_parent)
-
-                adapter = Group(new_parent, item)
-                if adapter and adapter.can_contain():
-                    adapter.group()
-
-                new_parent.request_update()
-        finally:
-            view.selection.set_dropzone_item(None)
-
-
-def apply_default_tool_set(view, event_manager, rubberband_state):
-    """The default tool set."""
-    view.add_controller(hover_tool(view))
-    view.add_controller(item_tool(view))
-    # chain.append(TextEditTool(view))
-    view.add_controller(rubberband_tool(view, rubberband_state))
